CL-643: More code generalizing: SchemaDao
