+++ mdm/common/src/main/java/com/gentics/mesh/search/IndexHandlerRegistry.java
@@ -3,7 +3,7 @@ package com.gentics.mesh.search;
 import java.util.List;
 
 import com.gentics.mesh.core.data.search.IndexHandler;
-import com.gentics.mesh.search.index.node.NodeIndexHandler;
+import com.gentics.mesh.core.search.index.node.NodeIndexHandler;
 
 /**
  * Registry for index handlers keep track of various listed handlers.
