+++ mdm/orientdb-api/src/main/java/com/gentics/mesh/core/data/schema/SchemaVersion.java
@@ -5,20 +5,11 @@ import static com.gentics.mesh.core.rest.MeshEvent.SCHEMA_CREATED;
 import static com.gentics.mesh.core.rest.MeshEvent.SCHEMA_DELETED;
 import static com.gentics.mesh.core.rest.MeshEvent.SCHEMA_UPDATED;
 
-import java.util.Iterator;
-import java.util.stream.Stream;
-
 import com.gentics.mesh.core.TypeInfo;
-import com.gentics.mesh.core.data.Bucket;
-import com.gentics.mesh.core.data.NodeGraphFieldContainer;
 import com.gentics.mesh.core.rest.schema.SchemaReference;
 import com.gentics.mesh.core.rest.schema.SchemaVersionModel;
 import com.gentics.mesh.core.rest.schema.impl.SchemaResponse;
 
-/**
- * Each schema update is stored within a dedicated schema container version in order to be able to keep track of changes in between different schema container
- * versions.
- */
 public interface SchemaVersion
 		extends GraphFieldSchemaContainerVersion<SchemaResponse, SchemaVersionModel, SchemaReference, HibSchemaVersion, HibSchema>, HibSchemaVersion {
 
@@ -28,36 +19,4 @@ public interface SchemaVersion
 	default TypeInfo getTypeInfo() {
 		return TYPE_INFO;
 	}
-
-	/**
-	 * Return a stream for {@link NodeGraphFieldContainer}'s that use this schema version and are versions for the given branch.
-	 *
-	 * @param branchUuid
-	 *            branch Uuid
-	 * @return
-	 */
-	Stream<? extends NodeGraphFieldContainer> getFieldContainers(String branchUuid);
-
-	/**
-	 * Return a stream for {@link NodeGraphFieldContainer}'s that use this schema version, are versions of the given branch and are listed within the given bucket.
-	 * @param branchUuid
-	 * @param bucket
-	 * @return
-	 */
-	Stream<? extends NodeGraphFieldContainer> getFieldContainers(String branchUuid, Bucket bucket);
-
-	/**
-	 * Returns an iterator for those {@link NodeGraphFieldContainer}'s which can be edited by users. Those are draft and publish versions.
-	 *
-	 * @param branchUuid Branch Uuid
-	 * @return
-	 */
-	Iterator<? extends NodeGraphFieldContainer> getDraftFieldContainers(String branchUuid);
-
-	/**
-	 * Check whether versioning is disabled by default or via the schema setting.
-	 * @return
-	 */
-	boolean isAutoPurgeEnabled();
-
 }
