+++ tests/tests-core/src/main/java/com/gentics/mesh/search/MicroschemaMappingTest.java
@@ -29,9 +29,9 @@ import com.gentics.mesh.core.rest.node.field.Field;
 import com.gentics.mesh.core.rest.schema.FieldSchema;
 import com.gentics.mesh.core.rest.schema.impl.SchemaCreateRequest;
 import com.gentics.mesh.core.rest.schema.impl.SchemaResponse;
+import com.gentics.mesh.core.search.index.node.NodeContainerMappingProvider;
+import com.gentics.mesh.core.search.index.node.NodeIndexHandler;
 import com.gentics.mesh.etc.config.search.ComplianceMode;
-import com.gentics.mesh.search.index.node.NodeContainerMappingProvider;
-import com.gentics.mesh.search.index.node.NodeIndexHandler;
 import com.gentics.mesh.test.MeshTestSetting;
 import com.gentics.mesh.test.TestSize;
 import com.gentics.mesh.test.context.AbstractMeshTest;
